Capstone Sytensis Prototype/BuildAPC/buildapc_project/buildapc_project/asgi.py
from channels.auth import AuthMiddlewareStack
from channels.routing import ProtocolTypeRouter, URLRouter
import your_app.routing  # assuming you have a routing.py in your app for WebSocket handling

# Setup the ASGI application to handle WebSocket connections
application = ProtocolTypeRouter({
    "http": get_asgi_application(),
    "websocket": AuthMiddlewareStack(
        URLRouter(
            your_app.routing.websocket_urlpatterns
        )
    ),
})
